pollution chatbot/backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
pymongo>=4.0.0
python-dotenv>=1.0.0
google-generativeai>=0.3.0
